Restore SessionLocal import
