+++ modules/inventory/forms.py
@@ -17,6 +17,10 @@ __all__ = [
     "LocationCreate",
     "LocationRead",
     "LocationUpdate",
+    "TagBase",
+    "TagCreate",
+    "TagRead",
+    "TagUpdate",
 ]
 
 
@@ -115,6 +119,28 @@ class LocationRead(LocationBase):
 class LocationUpdate(BaseModel):
     name: str | None = None
     location_type: str | None = None
+    version: int | None = Field(None, ge=1)
+    conflict_data: list[ConflictEntry] | None = None
+
+
+class TagBase(BaseSchema):
+    name: str
+
+
+class TagCreate(BaseModel):
+    name: str
+    version: int = Field(1, ge=1)
+    conflict_data: list[ConflictEntry] | None = None
+
+
+class TagRead(TagBase):
+    id: int
+
+    model_config = ConfigDict(from_attributes=True)
+
+
+class TagUpdate(BaseModel):
+    name: str | None = None
     version: int | None = Field(None, ge=1)
     conflict_data: list[ConflictEntry] | None = None
